+++ src/widgets/TreeWidget.hpp
@@ -67,7 +67,7 @@ public:
     void retranslateStrings();
 
 protected:
-    virtual void keyPressEvent( QKeyEvent *k );
+    virtual void keyPressEvent(QKeyEvent *event);
     virtual void dropEvent(QDropEvent *event);
     virtual void dragEnterEvent(QDragEnterEvent *event);
     virtual void dragMoveEvent(QDragMoveEvent *event);